+++ database/migrations/2022_04_05_095145_create_followups_table.php
@@ -15,7 +15,7 @@ class CreateFollowupsTable extends Migration
     {
         Schema::create('followups', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('report_id');
+            $table->integer('report_id');
             $table->text("building");
             $table->text("shift");
             $table->longText("why");
